TSeries

TimeSeriesEcon.TSeries — Type

struct TSeries{Frequency} <: AbstractVector{Float64}

Data structure representing a time-series vector. The following operations are allowed:

indexing using MIT (aka "moment-in-time") and UnitRange{MIT}

assignment using MIT and UnitRange{MIT}

In addition, most of the operations available to Julia vectors (+, -, *, etc.) are supported by TSeries as well.

Examples

Create TSeries

julia> x = TSeries(qq(2020, 1), ones(4))
TSeries{Quarterly} of length 4
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0

Index into TSeries

julia> x[qq(2020, 1)]
1.0

julia> x[qq(2020, 1):qq(2020, 2)]
TSeries{Quarterly} of length 2
2020Q1: 1.0
2020Q2: 1.0

Assignment using MIT

julia> x[qq(2020, 1)] = 100; x
TSeries{Quarterly} of length 4
2020Q1: 100.0
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0

julia> x[qq(2020, 1):qq(2020, 2)] = 100; x
TSeries{Quarterly} of length 4
2020Q1: 100.0
2020Q2: 100.0
2020Q3: 1.0
2020Q4: 1.0

Arithmetic Operations on TSeries

julia> x = TSeries(qq(2020, 1), ones(4))
TSeries{Quarterly} of length 4
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0

julia> 2*x + 98
TSeries{Quarterly} of length 4
2020Q1: 100.0
2020Q2: 100.0
2020Q3: 100.0
2020Q4: 100.0

julia> log(exp(x))
TSeries{Quarterly} of length 4
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0

TimeSeriesEcon.mitrange — Function

mitrange(x::TSeries)

Return an UnitRange{MIT{<:Frequency}} associated with x.

Examples

julia> mitrange(TSeries(qq(2020, 1), ones(4)))
2020Q1:2020Q4

TimeSeriesEcon.firstdate — Function

firstdate(x::TSeries)

Return an MIT indicating the first date in the TSeries.

Examples

julia> firstdate(TSeries(qq(2020, 1), ones(10)))
2020Q1

TimeSeriesEcon.lastdate — Function

lastdate(x::TSeries)

Return an MIT indicating the last date in the TSeries.

Examples

julia> lastdate(TSeries(qq(2020, 1), ones(10)))
2022Q2

TimeSeriesEcon.ppy — Function

ppy(::Frequency)

Returns the number of periods per year for a given Frequency, MIT, and TSeries

Examples

julia> ppy(Quarterly)                   # Frequency
4

julia> ppy(mm(2020, 1))                 # MIT
12

julia> ppy(TSeries(yy(2020), ones(3)))   # TSeries
1

ppy(::MIT)
ppy(::Type{MIT})

When applied to an MIT instance or type, return the ppy of its frequency.

ppy(::TSeries)
ppy(::Type{TSeries})

When applied to a TSeries instance or type, return the ppy of its frequency.

TimeSeriesEcon.shift — Function

shift(x::TSeries, n::Int64)

Shift dates of x back by k periods. Note: The implementation of is similar to IRIS ts{1}.

Examples

julia> shift(TSeries(qq(2020, 1), ones(4)), 1)
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0

julia> shift(TSeries(qq(2020, 1), ones(4)), -1)
TSeries{Quarterly} of length 4
2020Q2: 1.0
2020Q3: 1.0
2020Q4: 1.0
2021Q1: 1.0

See also: shift!

TimeSeriesEcon.shift! — Function

shift!(x::TSeries, n::Int64)

Shift dates of x back by k periods, in-place. Note: The implementation of is similar to IRIS ts{1}.

Examples

julia> x = TSeries(qq(2020, 1), ones(4));

julia> shift!(x, 1);

julia> x
TSeries{Quarterly} of length 4
2019Q4: 1.0
2020Q1: 1.0
2020Q2: 1.0
2020Q3: 1.0

See also: shift

TimeSeriesEcon.pct — Function

pct(x::TSeries, shift_value::Int64, islog::Bool)

Calculate percentage growth in x given a shift_value.

Note: The implementation is similar to IRIS.

Examples

julia> x = TSeries(yy(2000), Vector(1:4));

julia> pct(x, -1)
TSeries{Yearly} of length 3
2001Y: 100.0
2002Y: 50.0
2003Y: 33.33333333333333

See also: apct

TimeSeriesEcon.apct — Function

apct(x::TSeries, islog::Bool)

Calculate annualised percent rate of change in x.

Note: The implementation is similar to IRIS.

Examples

julia> x = TSeries(qq(2018, 1), Vector(1:8));

julia> apct(x)
TSeries{Quarterly} of length 7
2018Q2: 1500.0
2018Q3: 406.25
2018Q4: 216.04938271604937
2019Q1: 144.140625
2019Q2: 107.35999999999999
2019Q3: 85.26234567901243
2019Q4: 70.59558517284461

See also: pct

TimeSeriesEcon.nanrm! — Function

nanrm!(s::TSeries, type::Symbol)

Remove NaN values that are either at the beginning of the s and/or end of x.

Examples

julia> s = TSeries(yy(2018), [NaN, NaN, 1, 2, NaN]);

julia> nanrm!(s);

julia> s
TSeries{Yearly} of length 2
2020Y: 1.0
2021Y: 2.0

TimeSeriesEcon.@rec — Macro

@rec(eqn, rng)

Computes recursive calculations for the given eqn and rng.

Examples

julia> s = TSeries(ii(1), zeros(1));

julia> # Initial values

julia> s[ii(1)] = 0;

julia> s[ii(2)] = 1;

julia> @rec s[t] = s[t-1] + s[t-2] ii(3):ii(10)

julia> s
TSeries{Unit} of length 10
ii(1): 0.0
ii(2): 1.0
ii(3): 1.0
ii(4): 2.0
ii(5): 3.0
ii(6): 5.0
ii(7): 8.0
ii(8): 13.0
ii(9): 21.0
ii(10): 34.0